requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
Deprecated==1.3.1
distro==1.9.0
filelock==3.25.2
Flask==3.1.3
Flask-Limiter==4.1.1
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mlx==0.31.1
mlx-lm==0.31.1
mlx-metal==0.31.1
numpy==2.4.4
openai==2.30.0
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.27.2.2
pytest==9.0.2
PyYAML==6.0.3
regex==2026.3.32
rich==14.3.3
safetensors==0.7.0
sentencepiece==0.2.1
shellingham==1.5.4
sniffio==1.3.1
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.7
wrapt==2.1.2
